Add multi-source support to local and GCP provider
